Framework for UI implemented
Need to allow for flexibility.  Change hard-coding.

+++ Assets/Scripts/Items/InventoryItem.cs
@@ -3,5 +3,5 @@ using System.Collections;
 
 public class InventoryItem : MonoBehaviour {
     public string itemName;
-    public int currentAmount;
+    public float currentAmount;
 }
